+++ lib/screens/browser.dart
@@ -1,29 +1,19 @@
 import 'package:every_door/constants.dart';
 import 'package:every_door/helpers/circle_bounds.dart';
 import 'package:every_door/helpers/lifecycle.dart';
-import 'package:every_door/helpers/tile_layers.dart';
 import 'package:every_door/providers/api_status.dart';
 import 'package:every_door/providers/area.dart';
-import 'package:every_door/providers/changes.dart';
 import 'package:every_door/providers/editor_mode.dart';
-import 'package:every_door/providers/editor_settings.dart';
 import 'package:every_door/providers/geolocation.dart';
-import 'package:every_door/providers/imagery.dart';
 import 'package:every_door/providers/location.dart';
 import 'package:every_door/providers/need_update.dart';
 import 'package:every_door/providers/osm_api.dart';
-import 'package:every_door/providers/osm_auth.dart';
 import 'package:every_door/providers/osm_data.dart';
-import 'package:every_door/providers/poi_filter.dart';
 import 'package:every_door/screens/editor/map_chooser.dart';
-import 'package:every_door/screens/settings.dart';
-import 'package:every_door/screens/settings/account.dart';
 import 'package:every_door/screens/modes/entrances.dart';
-import 'package:every_door/widgets/filter.dart';
 import 'package:every_door/screens/modes/poi_list.dart';
+import 'package:every_door/widgets/navbar.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_dropdown_alert/alert_controller.dart';
-import 'package:flutter_dropdown_alert/model/data_alert.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:latlong2/latlong.dart' show LatLng;
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -80,39 +70,17 @@ class _BrowserPageState extends ConsumerState<BrowserPage> {
     }
   }
 
-  downloadAmenities(LatLng location) async {
+  downloadAmenities() async {
+    final location = ref.read(effectiveLocationProvider);
     final provider = ref.read(osmDataProvider);
     await provider.downloadAround(location);
     updateAreaStatus();
     ref.read(needMapUpdateProvider).trigger();
   }
 
-  uploadChanges(BuildContext context) async {
-    if (ref.read(authProvider) == null) {
-      Navigator.push(
-          context, MaterialPageRoute(builder: (context) => OsmAccountPage()));
-      return;
-    }
-
-    final loc = AppLocalizations.of(context)!;
-    try {
-      int count = await ref.read(osmApiProvider).uploadChanges(true);
-      AlertController.show(
-          loc.changesUploadedTitle,
-          loc.changesUploadedMessage(loc.changesCount(count)),
-          TypeAlert.success);
-    } on Exception catch (e) {
-      AlertController.show(
-          loc.changesUploadFailedTitle, e.toString(), TypeAlert.error);
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     final editorMode = ref.watch(editorModeProvider);
-    final apiStatus = ref.watch(apiStatusProvider);
-    final hasChangesToUpload = ref.watch(changesProvider).haveNoErrorChanges();
-    final hasFilter = ref.watch(poiFilterProvider).isNotEmpty;
 
     ref.listen(effectiveLocationProvider, (_, LatLng next) {
       updateAreaStatus();
@@ -134,23 +102,6 @@ class _BrowserPageState extends ConsumerState<BrowserPage> {
         break;
     }
 
-    final leftHand = ref.watch(editorSettingsProvider).leftHand;
-    final settingsButton = IconButton(
-      onPressed: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(builder: (context) => SettingsPage()),
-        );
-      },
-      icon: Icon(Icons.menu),
-    );
-    final modeButton = IconButton(
-      onPressed: () {
-        ref.read(editorModeProvider.notifier).next();
-      },
-      icon: Icon(kEditorModeIcons[editorMode]!),
-    );
-
     return WillPopScope(
       onWillPop: () async {
         if (ref.read(microZoomedInProvider) != null) {
@@ -164,81 +115,33 @@ class _BrowserPageState extends ConsumerState<BrowserPage> {
         }
       },
       child: Scaffold(
-        appBar: AppBar(
-          title: Text(kAppTitle, overflow: TextOverflow.fade),
-          leading: leftHand ? modeButton : settingsButton,
-          actions: [
-            if (!hasChangesToUpload)
-              IconButton(
-                onPressed: apiStatus != ApiStatus.idle
-                    ? null
-                    : () {
-                        final location = ref.read(effectiveLocationProvider);
-                        downloadAmenities(location);
-                      },
-                icon: Icon(Icons.download),
-              ),
-            if (hasChangesToUpload)
-              IconButton(
-                onPressed: apiStatus != ApiStatus.idle
-                    ? null
-                    : () async {
-                        uploadChanges(context);
-                      },
-                icon: Icon(Icons.upload, color: Colors.yellowAccent),
-              ),
-            IconButton(
-              onPressed: () {
-                setState(() {
-                  ref.read(selectedImageryProvider.notifier).toggle();
-                });
-              },
-              icon: Icon(ref.watch(selectedImageryProvider) == kOSMImagery
-                  ? Icons.map_outlined
-                  : Icons.map),
-            ),
-            if (editorMode == EditorMode.poi)
-              IconButton(
-                onPressed: () {
-                  showModalBottomSheet(
-                    context: context,
-                    builder: (BuildContext context) {
-                      return Container(
-                        color: Colors.white,
-                        height: 250.0,
-                        padding: EdgeInsets.all(15.0),
-                        child: PoiFilterPane(),
-                      );
-                    },
-                  );
-                },
-                icon: Icon(
-                  hasFilter ? Icons.filter_alt : Icons.filter_alt_outlined,
-                  color: hasFilter ? Colors.yellowAccent : null,
-                ),
-              ),
-            !leftHand ? modeButton : settingsButton,
+        body: Column(
+          children: [
+            Expanded(child: editorPanel),
+            BrowserNavigationBar(downloadAmenities: downloadAmenities),
           ],
         ),
-        body: editorPanel,
         floatingActionButton: editorMode == EditorMode.poi ||
                 editorMode == EditorMode.micromapping
-            ? FloatingActionButton(
-                child: Icon(Icons.add),
-                onPressed: () {
-                  ref.read(microZoomedInProvider.state).state = null;
-                  final location = ref.read(effectiveLocationProvider);
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => MapChooserPage(
-                        creating: true,
-                        location: location,
-                        closer: editorMode == EditorMode.micromapping,
+            ? Padding(
+                padding: const EdgeInsets.only(bottom: 50.0),
+                child: FloatingActionButton(
+                  child: Icon(Icons.add),
+                  onPressed: () {
+                    ref.read(microZoomedInProvider.state).state = null;
+                    final location = ref.read(effectiveLocationProvider);
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => MapChooserPage(
+                          creating: true,
+                          location: location,
+                          closer: editorMode == EditorMode.micromapping,
+                        ),
                       ),
-                    ),
-                  );
-                },
+                    );
+                  },
+                ),
               )
             : null,
       ),
@@ -271,7 +174,7 @@ class _BrowserPageState extends ConsumerState<BrowserPage> {
           ),
         ),
         onTap: () {
-          downloadAmenities(ref.read(effectiveLocationProvider));
+          downloadAmenities();
         },
       );
     return null;